utils: Add fetchWithCache()
